rendering  issue on transaction page, modal title
