+++ packages/components/built/SmartObjectFunction.d.ts
@@ -1,0 +1,4 @@
+export declare const getErrorMessage: (error: any) => string;
+export declare const getFnParamNames: (fn: string) => string[];
+export declare const getValueForType: (type: string, stringValue: string) => string | number | true | null | undefined;
+export declare const SmartObjectFunction: ({ computer, smartObject, functionsExist, options, setFunctionResult, setShow, setModalTitle, }: any) => import("react/jsx-runtime").JSX.Element;
