+++ packages/components/built/common/TypeSelectionDropdown.d.ts
@@ -1,0 +1,1 @@
+export declare const TypeSelectionDropdown: ({ id, onSelectMethod, dropdownList, selectedType }: any) => import("react/jsx-runtime").JSX.Element;
